+++ packages/preact/src/adapters/zodForm.ts
@@ -18,14 +18,15 @@ export function zodForm<TFieldValues extends FieldValues>(
 ): ValidateForm<TFieldValues> {
   return (values: PartialValues<TFieldValues>) => {
     const result = schema.safeParse(values);
-    return result.success
-      ? {}
-      : (result.error.issues.reduce<any>((errors, error) => {
-          const path = error.path.join('.');
-          if (!errors[path]) {
-            errors[path] = error.message;
-          }
-          return errors;
-        }, {}) as FormErrors<TFieldValues>);
+    const formErrors: Record<string, string> = {};
+    if (!result.success) {
+      for (const issue of result.error.issues) {
+        const path = issue.path.join('.');
+        if (!formErrors[path]) {
+          formErrors[path] = issue.message;
+        }
+      }
+    }
+    return formErrors as FormErrors<TFieldValues>;
   };
 }