+++ packages/qwik/src/adapters/valiField$.ts
@@ -1,23 +1,24 @@
 import { $, implicit$FirstArg, type QRL } from '@builder.io/qwik';
-import type { BaseSchema, BaseSchemaAsync } from 'valibot';
+import {
+  type GenericSchema,
+  type GenericSchemaAsync,
+  safeParseAsync,
+} from 'valibot';
 import type { FieldValue, MaybeFunction, ValidateField, Maybe } from '../types';
 
 /**
  * See {@link valiField$}
  */
 export function valiFieldQrl<TFieldValue extends FieldValue>(
-  schema: QRL<
-    MaybeFunction<
-      BaseSchema<TFieldValue, any> | BaseSchemaAsync<TFieldValue, any>
-    >
-  >
+  schema: QRL<MaybeFunction<GenericSchema | GenericSchemaAsync>>
 ): QRL<ValidateField<TFieldValue>> {
   return $(async (value: Maybe<TFieldValue>) => {
     const resolvedSchema = await schema.resolve();
-    const result = await (typeof resolvedSchema === 'function'
-      ? resolvedSchema()
-      : resolvedSchema
-    )._parse(value, { abortPipeEarly: true });
+    const result = await safeParseAsync(
+      typeof resolvedSchema === 'function' ? resolvedSchema() : resolvedSchema,
+      value,
+      { abortPipeEarly: true }
+    );
     return result.issues?.[0].message || '';
   });
 }