+++ packages/qwik/src/adapters/valiForm$.ts
@@ -1,5 +1,10 @@
 import { $, implicit$FirstArg, type QRL } from '@builder.io/qwik';
-import type { BaseSchema, BaseSchemaAsync } from 'valibot';
+import {
+  type GenericSchema,
+  type GenericSchemaAsync,
+  getDotPath,
+  safeParseAsync,
+} from 'valibot';
 import type {
   FieldValues,
   MaybeFunction,
@@ -12,27 +17,22 @@ import type {
  * See {@link valiForm$}
  */
 export function valiFormQrl<TFieldValues extends FieldValues>(
-  schema: QRL<
-    MaybeFunction<
-      BaseSchema<TFieldValues, any> | BaseSchemaAsync<TFieldValues, any>
-    >
-  >
+  schema: QRL<MaybeFunction<GenericSchema | GenericSchemaAsync>>
 ): QRL<ValidateForm<TFieldValues>> {
   return $(async (values: PartialValues<TFieldValues>) => {
     const resolvedSchema = await schema.resolve();
-    const result = await (typeof resolvedSchema === 'function'
-      ? resolvedSchema()
-      : resolvedSchema
-    )._parse(values, { abortPipeEarly: true });
-    return result.issues
-      ? result.issues.reduce<FormErrors<TFieldValues>>(
-          (errors, issue) => ({
-            ...errors,
-            [issue.path!.map(({ key }) => key).join('.')]: issue.message,
-          }),
-          {}
-        )
-      : ({} as FormErrors<TFieldValues>);
+    const result = await safeParseAsync(
+      typeof resolvedSchema === 'function' ? resolvedSchema() : resolvedSchema,
+      values,
+      { abortPipeEarly: true }
+    );
+    const formErrors: Record<string, string> = {};
+    if (result.issues) {
+      for (const issue of result.issues) {
+        formErrors[getDotPath(issue)!] = issue.message;
+      }
+    }
+    return formErrors as FormErrors<TFieldValues>;
   });
 }