+++ playgrounds/qwik/src/routes/actions/login/index.tsx
@@ -14,17 +14,19 @@ import * as v from 'valibot';
 import { FormHeader, TextInput, FormFooter, Response } from '~/components';
 
 const LoginSchema = v.object({
-  email: v.string([
-    v.minLength(1, 'Please enter your email.'),
-    v.email('The email address is badly formatted.'),
-  ]),
-  password: v.string([
-    v.minLength(1, 'Please enter your password.'),
-    v.minLength(8, 'You password must have 8 characters or more.'),
-  ]),
+  email: v.pipe(
+    v.string(),
+    v.nonEmpty('Please enter your email.'),
+    v.email('The email address is badly formatted.')
+  ),
+  password: v.pipe(
+    v.string(),
+    v.nonEmpty('Please enter your password.'),
+    v.minLength(8, 'You password must have 8 characters or more.')
+  ),
 });
 
-type LoginForm = v.Input<typeof LoginSchema>;
+type LoginForm = v.InferInput<typeof LoginSchema>;
 
 const getInitFormValues = (): InitialValues<LoginForm> => ({
   email: '',