+++ packages/qwik/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## v0.25.0 (June 06, 2024)
+
+- Upgrade `valiField$` and `valiForm$` adapter for Valibot v0.31.0
+- Improve `zodForm$` adapter
+
 ## v0.24.0 (April 30, 2024)
 
 - Replace `getFormDataValues` util with `decode-formdata` package